feat: Add navigator
Add a navigator in the bottom left hand corner. This serves two
functions:
1. It visualizes which way the user is facing on the horizontal axis.
2. It allows users to pan around using their keyboard only. This can be
   done by focusing one of the four directional buttons. Or by using the
   WASD keyboard shorcuts.

As part of this work, the styles are no longer scoped to the .video-js
prefix selector as that's not always present on videojs rendered videos
(e.g. when using it in a web application with custom handling).